refactor: change directory name
